Feat #120 - Get previous submissions paginated (#121)
* get submission by id

* retrieve submissions by category

* get submission by organization

* relocate submission unit tests
