+++ apps/server/swagger/submission-api.yml
@@ -2,7 +2,7 @@
 
 /submission/{submissionId}:
   get:
-    summary: Get active submission by ID
+    summary: Get submission by ID
     tags:
       - Submission
     parameters:
@@ -12,11 +12,11 @@
         required: true
     responses:
       200:
-        description: Active Submission
+        description: Submission details
         content:
           application/json:
             schema:
-              $ref: '#/components/schemas/ActiveSubmissionResult'
+              $ref: '#/components/schemas/SubmissionResult'
 
       401:
         $ref: '#/components/responses/UnauthorizedError'
@@ -42,7 +42,7 @@
         content:
           application/json:
             schema:
-              $ref: '#/components/schemas/ActiveSubmissionResult'
+              $ref: '#/components/schemas/SubmissionResult'
       400:
         $ref: '#/components/responses/BadRequest'
       401:
@@ -91,7 +91,7 @@
         content:
           application/json:
             schema:
-              $ref: '#/components/schemas/ActiveSubmissionResult'
+              $ref: '#/components/schemas/SubmissionResult'
       400:
         $ref: '#/components/responses/BadRequest'
       401:
@@ -107,18 +107,22 @@
 
 /submission/category/{categoryId}:
   get:
-    summary: Retrieve the Active Submissions for a category in this user session
+    summary: Retrieve the Submissions for a category in this user session
     tags:
       - Submission
     parameters:
       - $ref: '#/components/parameters/path/CategoryId'
+      - $ref: '#/components/parameters/query/Page'
+      - $ref: '#/components/parameters/query/PageSize'
+      - $ref: '#/components/parameters/query/OnlyActive'
+      - $ref: '#/components/parameters/query/Organization'
     responses:
       200:
-        description: Active Submissions
+        description: Submissions details
         content:
           application/json:
             schema:
-              $ref: '#/components/schemas/ActiveSubmissionsSummaryResult'
+              $ref: '#/components/schemas/SubmissionsSummaryResult'
 
       401:
         $ref: '#/components/responses/UnauthorizedError'
@@ -132,6 +136,7 @@
 /submission/category/{categoryId}/organization/{organization}:
   get:
     summary: Retrieve the Active Submission for a category and organization in this user session
+    deprecated: true
     tags:
       - Submission
     parameters:
