+++ packages/data-provider/src/services/submission/submission.ts
@@ -13,18 +13,19 @@ import {
 	canTransitionToClosed,
 	checkEntityFieldNames,
 	checkFileNames,
-	parseActiveSubmissionResponse,
-	parseActiveSubmissionSummaryResponse,
+	parseSubmissionResponse,
+	parseSubmissionSummaryResponse,
 	removeItemsFromSubmission,
 } from '../../utils/submissionUtils.js';
 import {
-	ActiveSubmissionSummaryResponse,
 	CommitSubmissionResult,
 	CREATE_SUBMISSION_STATUS,
 	type CreateSubmissionResult,
+	type PaginationOptions,
 	SUBMISSION_ACTION_TYPE,
 	SUBMISSION_STATUS,
 	type SubmissionActionType,
+	SubmissionSummaryResponse,
 } from '../../utils/types.js';
 import processor from './processor.js';
 
@@ -232,43 +233,64 @@ const service = (dependencies: BaseDependencies) => {
 	};
 
 	/**
-	 * Get an active Submission by Category
-	 * @param {Object} params
-	 * @param {number} params.categoryId
-	 * @param {string} params.userName
-	 * @returns  One Active Submission
+	 * Get Submissions by Category
+	 * @param {number} categoryId - The ID of the category for which data is being fetched.
+	 * @param {Object} paginationOptions - Pagination properties
+	 * @param {number} paginationOptions.page - Page number
+	 * @param {number} paginationOptions.pageSize - Items per page
+	 * @param {Object} filterOptions
+	 * @param {boolean} filterOptions.onlyActive - Filter by Active status
+	 * @param {string} filterOptions.userName - User Name
+	 * @returns an array of Submission
 	 */
-	const getActiveSubmissionsByCategory = async ({
-		categoryId,
-		userName,
-	}: {
-		categoryId: number;
-		userName: string;
-	}): Promise<ActiveSubmissionSummaryResponse[] | undefined> => {
-		const { getActiveSubmissionsWithRelationsByCategory } = submissionRepository(dependencies);
 
-		const submissions = await getActiveSubmissionsWithRelationsByCategory({ userName, categoryId });
-		if (!submissions || submissions.length === 0) {
-			return;
+	const getSubmissionsByCategory = async (
+		categoryId: number,
+		paginationOptions: PaginationOptions,
+		filterOptions: {
+			onlyActive: boolean;
+			userName: string;
+			organization?: string;
+		},
+	): Promise<{
+		result: SubmissionSummaryResponse[];
+		metadata: { totalRecords: number; errorMessage?: string };
+	}> => {
+		const { getSubmissionsWithRelationsByCategory, getTotalSubmissionsByCategory } = submissionRepository(dependencies);
+
+		const recordsPaginated = await getSubmissionsWithRelationsByCategory(categoryId, paginationOptions, filterOptions);
+		if (!recordsPaginated || recordsPaginated.length === 0) {
+			return {
+				result: [],
+				metadata: {
+					totalRecords: 0,
+				},
+			};
 		}
 
-		return submissions.map((response) => parseActiveSubmissionSummaryResponse(response));
+		const totalRecords = await getTotalSubmissionsByCategory(categoryId, filterOptions);
+		return {
+			metadata: {
+				totalRecords,
+			},
+			result: recordsPaginated.map((response) => parseSubmissionSummaryResponse(response)),
+		};
 	};
 
 	/**
-	 * Get Active Submission by Submission ID
+	 * Get Submission by Submission ID
 	 * @param {number} submissionId A Submission ID
-	 * @returns One Active Submission
+	 * @returns One Submission
 	 */
-	const getActiveSubmissionById = async (submissionId: number) => {
-		const { getActiveSubmissionWithRelationsById } = submissionRepository(dependencies);
+	const getSubmissionById = async (submissionId: number) => {
+		const { getSubmissionWithRelationsById } = submissionRepository(dependencies);
 
-		const submission = await getActiveSubmissionWithRelationsById(submissionId);
+		const submission = await getSubmissionWithRelationsById(submissionId);
 		if (_.isEmpty(submission)) {
 			return;
 		}
 
-		return parseActiveSubmissionResponse(submission);
+		return parseSubmissionResponse(submission);
 	};
 
 	/**
@@ -287,7 +309,7 @@ const service = (dependencies: BaseDependencies) => {
 		categoryId: number;
 		userName: string;
 		organization: string;
-	}): Promise<ActiveSubmissionSummaryResponse | undefined> => {
+	}): Promise<SubmissionSummaryResponse | undefined> => {
 		const { getActiveSubmissionWithRelationsByOrganization } = submissionRepository(dependencies);
 
 		const submission = await getActiveSubmissionWithRelationsByOrganization({ organization, userName, categoryId });
@@ -295,7 +317,7 @@ const service = (dependencies: BaseDependencies) => {
 			return;
 		}
 
-		return parseActiveSubmissionSummaryResponse(submission);
+		return parseSubmissionSummaryResponse(submission);
 	};
 
 	/**
@@ -444,8 +466,8 @@ const service = (dependencies: BaseDependencies) => {
 		commitSubmission,
 		deleteActiveSubmissionById,
 		deleteActiveSubmissionEntity,
-		getActiveSubmissionsByCategory,
-		getActiveSubmissionById,
+		getSubmissionsByCategory,
+		getSubmissionById,
 		getActiveSubmissionByOrganization,
 		getOrCreateActiveSubmission,
 		uploadSubmission,
