+++ packages/data-provider/src/utils/types.ts
@@ -196,16 +196,16 @@ export type PaginationOptions = {
 	pageSize: number;
 };
 
-export type DataInsertsActiveSubmissionSummary = {
+export type DataInsertsSubmissionSummary = {
 	batchName: string;
 	recordsCount: number;
 };
 
-export type DataUpdatesActiveSubmissionSummary = {
+export type DataUpdatesSubmissionSummary = {
 	recordsCount: number;
 };
 
-export type DataDeletesActiveSubmissionSummary = {
+export type DataDeletesSubmissionSummary = {
 	recordsCount: number;
 };
 
@@ -220,9 +220,9 @@ export type CategoryActiveSubmission = {
 };
 
 /**
- * Response type for Get Active Submission by Submission ID endpoint
+ * Response type for Get Submission by Submission ID endpoint
  */
-export type ActiveSubmissionResponse = {
+export type SubmissionResponse = {
 	id: number;
 	data: SubmissionData;
 	dictionary: DictionaryActiveSubmission;
@@ -237,21 +237,21 @@ export type ActiveSubmissionResponse = {
 };
 
 /**
- * Response type of Get Active Submission by Organization Endpoint
+ * Response type of Get Submission by Organization Endpoint
  * override 'data' object to contain a summary of records
  */
-export type ActiveSubmissionSummaryResponse = Omit<ActiveSubmissionResponse, 'data'> & {
+export type SubmissionSummaryResponse = Omit<SubmissionResponse, 'data'> & {
 	data: {
-		inserts?: Record<string, DataInsertsActiveSubmissionSummary>;
-		updates?: Record<string, DataUpdatesActiveSubmissionSummary>;
-		deletes?: Record<string, DataDeletesActiveSubmissionSummary>;
+		inserts?: Record<string, DataInsertsSubmissionSummary>;
+		updates?: Record<string, DataUpdatesSubmissionSummary>;
+		deletes?: Record<string, DataDeletesSubmissionSummary>;
 	};
 };
 
 /**
- * Retrieve Active Submission object from repository
+ * Retrieve Submission object from repository
  */
-export type ActiveSubmissionSummaryRepository = {
+export type SubmissionSummaryRepository = {
 	id: number;
 	data: SubmissionData;
 	dictionary: Pick<Dictionary, 'name' | 'version'>;
